+++ iridauploader/core/upload.py
@@ -17,7 +17,7 @@ from iridauploader.model import DirectoryStatus
 from . import api_handler, parsing_handler, logger, exit_return, upload_helpers
 
 
-def upload_run_single_entry(directory, force_upload=False, upload_mode=None):
+def upload_run_single_entry(directory, force_upload=False, upload_mode=None, continue_upload=False):
     """
     This function acts as a single point of entry for uploading a directory
 
@@ -26,20 +26,16 @@ def upload_run_single_entry(directory, force_upload=False, upload_mode=None):
     :param directory: Directory of the sequencing run to upload
     :param force_upload: When set to true, the upload status file will be ignored and file will attempt to be uploaded
     :param upload_mode: String with upload mode to use. When None, default is used.
+    :param continue_upload: When set, a PARTIAL status run will be continued from where it left off.
     :return: ExitReturn
     """
 
     directory_status = parsing_handler.get_run_status(directory)
+    parse_as_partial = False
 
     # Check that directory is writeable, or readonly mode is enabled
     if upload_helpers.directory_has_readonly_conflict(directory_status.directory):
         error_msg = 'Directory cannot be written to. Please check permissions or use readonly mode'
-        logging.error(error_msg)
-        return exit_error(error_msg)
-    # Check if a run is invalid, an invalid run cannot be uploaded.
-    elif directory_status.status_equals(DirectoryStatus.INVALID):
-        error_msg = "ERROR! Run in directory {} is invalid. Returned with message: '{}'".format(
-            directory_status.directory, directory_status.message)
         logging.error(error_msg)
         return exit_error(error_msg)
     # Check if run is New or Delayed, and then do delay logic
@@ -53,9 +49,29 @@ def upload_run_single_entry(directory, force_upload=False, upload_mode=None):
         else:
             logging.debug("Run is delayed, exiting")
             return exit_success()
+    # Check if run is partial, if we are continuing partial runs this block will enable parsing as partial
+    # NOTE: its assumed that at most only one of `continue_upload` and `force_upload` are set
+    elif directory_status.status_equals(DirectoryStatus.PARTIAL):
+        if continue_upload:
+            # Happy path for continuing an upload
+            logging.info("Continuing upload on a partial run.")
+            parse_as_partial = True
+        elif force_upload:
+            # Note: This is "happy path" 2, where upload starts from beginning with force
+            logging.info("Run with status {} is being force uploaded".format(directory_status.status))
+        else:
+            error_msg = "ERROR! Directory status is PARTIAL. This run can be continued with the --continue_partial" \
+                        " argument, or restarted from the beginning with the --force argument."
+            logging.error(error_msg)
+            return exit_error(error_msg)
+    # Check if a run is invalid, an invalid run cannot be uploaded.
+    elif directory_status.status_equals(DirectoryStatus.INVALID):
+        error_msg = "ERROR! Run in directory {} is invalid. Returned with message: '{}'".format(
+            directory_status.directory, directory_status.message)
+        logging.error(error_msg)
+        return exit_error(error_msg)
     # Check if run is any other status, if force upload is set, continue, otherwise exit
     elif (directory_status.status_equals(DirectoryStatus.ERROR)
-          or directory_status.status_equals(DirectoryStatus.PARTIAL)
           or directory_status.status_equals(DirectoryStatus.COMPLETE)):
         if force_upload:
             # Note: This is "happy path" 2, where upload continues with force
@@ -74,10 +90,10 @@ def upload_run_single_entry(directory, force_upload=False, upload_mode=None):
         upload_mode = api_handler.get_default_upload_mode()
 
     # upload
-    return _validate_and_upload(directory_status, upload_mode)
+    return _validate_and_upload(directory_status, upload_mode, parse_as_partial)
 
 
-def batch_upload_single_entry(batch_directory, force_upload=False, upload_mode=None):
+def batch_upload_single_entry(batch_directory, force_upload=False, upload_mode=None, continue_upload=False):
     """
     This function acts as a single point of entry for batch uploading run directories
 
@@ -88,6 +104,7 @@ def batch_upload_single_entry(batch_directory, force_upload=False, upload_mode=N
     :param batch_directory: Directory containing sequencing run directories to upload
     :param force_upload: When set to true, the upload status file will be ignored and file will attempt to be uploaded
     :param upload_mode: String with upload mode to use. When None, default is used.
+    :param continue_upload: When True, continues uploading existing partial runs from where they left off
     :return: ExitReturn
     """
     logging.debug("batch_upload_single_entry:Starting {} with force={}".format(batch_directory, force_upload))
@@ -102,8 +119,12 @@ def batch_upload_single_entry(batch_directory, force_upload=False, upload_mode=N
                      "%30sDETAILS: %s"
                      % (directory_status.directory, "", directory_status.status, "", directory_status.message))
 
+    # upload_list contains dicts with the following format
+    # {'status': DirectoryStatus, 'partial': boolean}
     upload_list = []
+    # delayed_list contains DirectoryStatus objects
     delayed_list = []
+
     for directory_status in directory_status_list:
         logging.info("Analysing directory: {}".format(directory_status.directory))
         # ignore invalid directories
@@ -114,22 +135,34 @@ def batch_upload_single_entry(batch_directory, force_upload=False, upload_mode=N
               or directory_status.status_equals(DirectoryStatus.DELAYED)):
             if force_upload:
                 logging.debug("BATCH: Run is being added with force")
-                upload_list.append(directory_status)
+                upload_list.append({"status": directory_status, "partial": False})
             elif progress.run_is_ready_with_delay(directory_status):
                 # Note: This is the "happy path" where upload continues
                 logging.debug("BATCH: Run is ready to upload")
-                upload_list.append(directory_status)
+                upload_list.append({"status": directory_status, "partial": False})
             else:
                 logging.debug("BATCH: Run is delayed")
                 delayed_list.append(directory_status)
-
+        # Check if run is partial, if we are continuing partial runs this block will enable parsing as partial
+        # NOTE: its assumed that at most only one of `continue_upload` and `force_upload` are set
+        elif directory_status.status_equals(DirectoryStatus.PARTIAL):
+            if continue_upload:
+                # Happy path for continuing an upload
+                logging.debug("BATCH: Run is ready to continue upload as partial")
+                upload_list.append({"status": directory_status, "partial": True})
+            elif force_upload:
+                # Note: This is "happy path" 2, where upload continues with force
+                logging.debug("BATCH: Partial Run is being added with force")
+                upload_list.append({"status": directory_status, "partial": False})
+            else:
+                logging.debug("BATCH: Partial Run is skipped")
+                continue
         # other statuses are error, partial, and complete runs, force upload allows theses
         elif (directory_status.status_equals(DirectoryStatus.ERROR)
-              or directory_status.status_equals(DirectoryStatus.PARTIAL)
               or directory_status.status_equals(DirectoryStatus.COMPLETE)):
             if force_upload:
                 logging.debug("BATCH: Run is forced for upload")
-                upload_list.append(directory_status)
+                upload_list.append({"status": directory_status, "partial": False})
             else:
                 logging.debug("BATCH: Run is skipped")
                 continue
@@ -149,13 +182,12 @@ def batch_upload_single_entry(batch_directory, force_upload=False, upload_mode=N
     if upload_mode is None:
         upload_mode = api_handler.get_default_upload_mode()
 
-    # run upload, keep track of which directories did not upload
+    # run uploads (including partial), keep track of which directories did not upload
     error_list = []
-    for directory_status in upload_list:
-        logging.info("Starting upload for {}".format(directory_status.directory))
-        result = _validate_and_upload(directory_status, upload_mode)
+    for directory_status_dict in upload_list:
+        result = _validate_and_upload(directory_status_dict['status'], upload_mode, directory_status_dict['partial'])
         if result.exit_code == exit_return.EXIT_CODE_ERROR:
-            error_list.append(directory_status)
+            error_list.append(directory_status_dict['status'])
 
     logging.info("Uploads completed with {} error(s)".format(len(error_list)))
     for directory_status in error_list:
@@ -166,7 +198,7 @@ def batch_upload_single_entry(batch_directory, force_upload=False, upload_mode=N
     return exit_success()
 
 
-def _validate_and_upload(directory_status, upload_mode):
+def _validate_and_upload(directory_status, upload_mode, continue_from_partial):
     """
     This function attempts to upload a single run directory
 
@@ -178,14 +210,17 @@ def _validate_and_upload(directory_status, upload_mode):
 
     :param directory_status: DirectoryStatus object that has directory to try upload
     :param upload_mode: String, mode to use when uploading assemblies
+    :param continue_from_partial: when set, already uploaded samples will be skipped, and existing run_id will be used
     :return: ExitReturn
     """
     logging_start_block(directory_status.directory)
     logging.debug("upload_run_single_entry:Starting {}".format(directory_status.directory))
+    logging.info("Starting upload for '{}' with continue partial upload = '{}'".format(
+        directory_status.directory, str(continue_from_partial)))
 
     try:
         # Starting upload process: Parse and do offline verification
-        sequencing_run = upload_helpers.parse_and_validate(directory_status)
+        sequencing_run = upload_helpers.parse_and_validate(directory_status, continue_from_partial)
         upload_helpers.verify_upload_mode(upload_mode)
         upload_helpers.init_file_status_list_from_sequencing_run(sequencing_run, directory_status)
 
@@ -194,7 +229,7 @@ def _validate_and_upload(directory_status, upload_mode):
         upload_helpers.irida_prep_and_validation(sequencing_run, directory_status)
 
         # Upload run
-        upload_helpers.upload_sequencing_run(sequencing_run, directory_status, upload_mode)
+        upload_helpers.upload_sequencing_run(sequencing_run, directory_status, upload_mode, continue_from_partial)
 
     except (progress.exceptions.DirectoryError,
             parsers.exceptions.ValidationError,