+++ iridauploader/gui/threads.py
@@ -4,8 +4,7 @@ import PyQt5.QtCore as QtCore
 
 from pprint import pformat
 
-from iridauploader.core import upload, parsing_handler, exit_return
-from iridauploader.config import config
+from iridauploader.core import upload, parsing_handler, exit_return, upload_helpers
 from iridauploader.parsers import exceptions
 
 
@@ -50,15 +49,17 @@ class ParseThread(QtCore.QThread):
     def __init__(self):
         super().__init__()
         self._directory = ""
+        self._parse_as_partial = False
         self._run = None
         self._error = None
 
-    def set_vars(self, directory):
+    def set_vars(self, directory, parse_as_partial):
         """
         Sets the variables in the object to the ones passed in
         :return:
         """
         self._directory = directory
+        self._parse_as_partial = parse_as_partial
 
     def get_run(self):
         return self._run
@@ -72,7 +73,12 @@ class ParseThread(QtCore.QThread):
         :return:
         """
         try:
-            self._run = parsing_handler.parse_and_validate(self._directory)
+            status = parsing_handler.get_run_status(self._directory)
+            seq_run = parsing_handler.parse_and_validate(self._directory)
+            if self._parse_as_partial:
+                seq_run = upload_helpers.set_uploaded_samples_to_skip(seq_run, status.get_sample_status_list())
+            self._run = seq_run
+
         except exceptions.DirectoryError as e:
             # Directory was not valid for some reason
             full_error = "GUI: ERROR! An error occurred with directory '{}', with message: {}".format(e.directory,
@@ -105,7 +111,7 @@ class ParseThread(QtCore.QThread):
             self._run = None
         except Exception as e:
             # Some other error occurred
-            full_error = "GUI: ERROR! An error occurred while parsing: {}".format(str(e))
+            full_error = "GUI: ERROR! An unknown error occurred while parsing: {}".format(str(e))
             logging.error(full_error)
             self._error = "ERROR! An error occurred while parsing: {}".format(str(e))
             self._run = None
@@ -123,29 +129,35 @@ class UploadThread(QtCore.QThread):
         self._run_dir = ""
         self._upload_mode = None
         self._exit_return = None
+        self._partial_continue = False
 
-    def set_vars(self, run_dir, upload_mode):
+    def set_vars(self, run_dir, upload_mode, partial_continue):
         """
         Sets the variables in the object to the ones passed in
         :param run_dir:
-        :param force_state:
-        :param upload_mode
+        :param upload_mode:
+        :param partial_continue:
         :return:
         """
         self._run_dir = run_dir
         self._upload_mode = upload_mode
+        self._partial_continue = partial_continue
 
     def run(self):
         """
         This runs when the threads start call is done
 
-        When uploading, it will always upload with force_upload=True,
-        s.t. we ignore delays and simplify earlier run parse logic
+        When uploading, the default is to upload with force_upload=True,
+          when continuing a partial upload, force_upload=False instead
+          s.t. we ignore delay logic and simplify earlier run parse logic
         :return:
         """
+
+        force_upload = not self._partial_continue
         self._exit_return = upload.upload_run_single_entry(directory=self._run_dir,
-                                                           force_upload=True,
-                                                           upload_mode=self._upload_mode)
+                                                           force_upload=force_upload,
+                                                           upload_mode=self._upload_mode,
+                                                           continue_upload=self._partial_continue)
         pass
 
     def is_success(self):